backend/requirements2.txt
fastapi==0.115.12
uvicorn==0.34.1
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.6
python-jose[cryptography]
bcrypt
passlib[bcrypt]
python-multipart>=0.0.6
pydantic==2.11.3
python-dotenv>=1.0.0
requests
httpx
docxtpl
weasyprint
boto3
supabase
openai==1.75.0
pypdf2==3.0.1
